requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
dnspython==2.6.1
googlesearch-python==1.2.5
idna==3.10
prettytable==3.11.0
python-dateutil==2.9.0.post0
python-whois==0.9.4
requests==2.32.3
six==1.16.0
soupsieve==2.6
tabulate==0.9.0
urllib3==2.2.3
wcwidth==0.2.13
